Fix bug for parameters not being considered
Pass 3: Added parameters into the correct methodscopes.
Pass 4: Checked parameters when looking for the declaration of a
variable to be captured

Small note: changed handle expression statements in pass 4 to not break
if the expression does not have a right hand side.

Might have to check if variables that need to be captured being used in some method calls.